Move DoProteinSearchAction
